telegram_bot/requirements.txt
flask
requests
telethon
python-dotenv
uvicorn
supervisor